Missing tests for new function

+++ src/UniGetUI.PackageEngine.Managers.Generic.NuGet/BaseNuGet.cs
@@ -44,7 +44,7 @@ namespace UniGetUI.PackageEngine.Managers.PowerShellManager
         private struct SearchResult
         {
             public string version;
-            public float version_float;
+            public double version_float;
             public string id;
         }
 
@@ -89,7 +89,7 @@ namespace UniGetUI.PackageEngine.Managers.PowerShellManager
 
                         string id = Regex.Match(match.Value, "Id='([^<>']+)'").Groups[1].Value;
                         string version = Regex.Match(match.Value, "Version='([^<>']+)'").Groups[1].Value;
-                        float float_version = CoreTools.GetVersionStringAsFloat(version);
+                        double float_version = CoreTools.GetVersionStringAsFloat(version);
                         Match title = Regex.Match(match.Value, "<title[ \\\"\\=A-Za-z0-9]+>([^<>]+)<\\/title>");
 
                         if (AlreadyProcessedPackages.ContainsKey(id) && AlreadyProcessedPackages[id].version_float >= float_version)
